Allow passing imagePullSecrets to deployment and statefulset
Prior to this commit users were able to pass custom images to the chart,
but usage of this feature was limited by using public registries, which
do not require authorization only.

This change adds possibility to pass imagePullSecrets name, thus
allowing usage of the chart with private registries. Users can define
one of

- zabbixServer.image.pullSecrets
- zabbixagent.image.pullSecrets
- zabbixproxy.image.pullSecrets
- zabbixweb.image.pullSecrets

to set corresponding imagePullSecrets for different components.

Intended usage is the following:

```console
$ cat values.yaml
...
zabbixServer:
  image:
    repository: registry.example.com/zabbix-custom/server
    pullSecrets:
      - server-registry-secret
...
zabbixweb:
  image:
    repository: registry.example.com/zabbix-custom/web
    pullSecrets:
      - web-registry-secret
```

Please note, that the secret in the cluster should be created prior to
chart deployment manually. The command from
https://kubernetes.io/docs/tasks/configure-pod-container/pull-image-private-registry/#registry-secret-existing-credentials
can be used in order to create corresponding secret in kubernetes
cluster.

Signed-off-by: Oleksandr Kovalchuk <anxolerd@outlook.com>