support DB init with multiple zabbix servers
When starting more than one Zabbix server containers with valid DB
connection credentials at the same time on an empty database, all of
them would try to connect and install the Zabbix DB schema to it at the
same time, leading to errors and invalid queries.

This commit implements an init container on the Zabbix Server pods that
not only waits for the DB to be accessible IP-wise, but also for the
schema to be available. This practically deactivates the schema
auto-creation for the Zabbix Server pods.

At the same time, a one-time Job has been added, which uses the Zabbix
Server image with a slightly modified docker-entrypoint.sh in order to
only wait for DB connection and then to install Zabbix DB schema, but
not to prepare config nor start Zabbix server.

With the combination of these two, we can now start off with a multinode
Zabbix HA cluster in Kubernetes.