fix: added additionally step to wait for db schema
if the step to create the db schema takes too long, the wait-job that
should prevent the Zabbix server container to start unless the database
is "really" ready, is leaving off too early and Zabbix server crashes at
start. This commit prevents this error by adding an additional step to
only start the Zabbix server container if there are more than 0 users in
the users table.