Minor improvement on jobs, version bump
Up to now, it was not possible to upgrade to a new major version of
Zabbix just by using `helm upgrade` because the job objects' "image"
setting is immutable, causing the job "init-db-schema" not to be able to
update and therefore failing the deployment.

Changed the job to disappear after 120 seconds after successful run
makes the major release upgrade possible.

Also set the same `ttlSecondsAfterFinished` to 24h for jobs created by
the cron job to automatically clean up orphaned HA nodes.

Made a version bump to Zabbix 6.0.10 and helm chart version 3.3.1.

Updated a comment regarding TimescaleDB support (Zabbix as of now
supports only TimescaleDB up to version 2.7.x).