new way to only init DB from 6.2.7 on
Since Zabbix 6.2.7, the zabbix-server-* container images have received a
new and partly rewritten docker-entrypoint.sh, and within this some
function names and other structures have changed, so that the old (quit
ugly, by the way) method of modifying the script with sed to only
initialize the DB schema does not work anymore. The good part of it: the
image's docker-entrypoint.sh does now support an "init_db_only"
attribute that does exactly what we want.

Therefore I have now changed the job manifest to first check which
version of Zabbix server is inside it, and in case it is newer than
6.2.6, call docker-entrypoint.sh with the new 'init_db_only' argument,
  and employ the old technique other case.