Add support for browser monitoring
This has been introduced into Zabbix 7.0 to support running Selenium based
website monitoring. For this to work, you need a running webdriver and web browser.
This is now integrated into the chart and can be enabled using --set zabbixBrowserMonitoring.enabled=true

Furthermore, it allows referencing a customWebDriverURL when an existing
Selenium grid is already set up and can be used.