remove service for Zabbix Agent sidecar
It makes no sense to have a service in such case, as these sidecar
containers are there solely to show up the built-in "Zabbix server"
host's interface availability to get green. This is done by using
"localhost", which no Service is needed for.

A Service indeed implements loadbalancing over all matched pods, which
in this case would lead to connections being round-robined between
Zabbix Agent containers on different Zabbix Server pods, if using Zabbix
Server HA, which is senseless.